Instancing implemented - Starting particle sim
At this point the rendering has been nearly completely taken care of,
and the priority is to start implementing the conversion from a particle
system to an instancing system.

Reflection: I'm actually very happy I got this far!! I'm pretty
confident I can finish this project - the amount of time is still
unknown depending on how pretty I want to make it. But overall I don't
feel that this will fail.